% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdRoit.est.loo.R
\name{AdRoit.est.loo}
\alias{AdRoit.est.loo}
\title{Estimate the proportions for a single sample which is left out in training data}
\usage{
AdRoit.est.loo(n, bulk.sample, single.ref, silent = FALSE)
}
\arguments{
\item{n}{index of sample (column) to be estimated.}

\item{bulk.sample}{a matrix or data.frame with the rows being genes and columns being samples (or spatial spots for spatial transcriptomics data).}

\item{single.ref}{the reference object built by the function `ref.build()`.}

\item{silent}{whether to print out messages. Default is FALSE.}
}
\value{
the cell type proportions for the nth sample.
}
\description{
This function estimate the proportions of a single bulk sample while being left out when estimating gene cross-sample variations and cross-platform biases.
}
