% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdRoit.est.R
\name{AdRoit.est}
\alias{AdRoit.est}
\title{Estimate the proportions of cell types given by the reference}
\usage{
AdRoit.est(bulk.sample, single.ref, silent = FALSE)
}
\arguments{
\item{bulk.sample}{a matrix or data.frame with the rows being genes and columns being samples (or spatial spots for spatial transcriptomics data).}

\item{single.ref}{the reference object built by the function `ref.build()`.}

\item{silent}{whether to print out messages. Default is FALSE.}
}
\value{
a matrix with rows being cell types and columns being bulk samples, and entries are estimated proportions.
}
\description{
This is the main function for estimating cell type proportions.
}
