% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BubblePlot.R
\name{BubblePlot}
\alias{BubblePlot}
\title{Make bubble plot to compare estimated proportions with truth}
\usage{
BubblePlot(prop.ests, prop.real, labels = c("Estimatd", "Truth"))
}
\arguments{
\item{prop.ests}{a matrix or data frame of estimated proporitons with row being cell types and columns being samples (spatial spots).}

\item{prop.real}{a matrix or data frame of the ground truth proportions.}

\item{labels}{a vector of two elements, labels the first two arguments. Default is `c("Estimated", "True")`.}
}
\value{
a ggplot object. Use `print` to render the plot.
}
\description{
This function makes overlaid bubble plots that help visualize the difference between estimated proportions and the true proportions.
}
