% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.Markers.AdRoit.R
\name{get.Markers.AdRoit}
\alias{get.Markers.AdRoit}
\title{Function to get top marker genes for each cell cluster using AdRoit simplified way}
\usage{
get.Markers.AdRoit(SC.counts, annotation)
}
\arguments{
\item{SC.counts}{the raw single cell count matrix, sparse matrix (dgCMatrix). The rows are genes and the columns are cells.}

\item{annotation}{a vector of length equal to number of cells, each element annotate the cell type of a cell in the count matrix.}
}
\value{
a matrix containing a ranked list of putative markers, and associated statistics (p-values, fold change, etc.)
}
\description{
This function first normalizes the single cell count matrix (gene by cell) by equalizing the total counts then take `log1p`,
and perform Wilcoxon test to identify top differential expressed genes in each cell type.
}
