% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.Markers.Seurat.R
\name{get.Markers.Seurat}
\alias{get.Markers.Seurat}
\title{Function to find marker genes per cell cluster using Seurat}
\usage{
get.Markers.Seurat(ref.obj, annotations)
}
\arguments{
\item{ref.obj}{a Seurat object after normalization and scaling.}

\item{annotations}{a vector of length equal to number of cells, each element annotates the cell type of a cell in the `ref.obj`.}
}
\value{
a matrix containing a ranked list of putative markers, and associated statistics (e.g., p-values, fold change, etc.)
}
\description{
This function is a wraper to find markers for cell clusters using Seruat
}
