% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.TopMarkers.R
\name{get.TopMarkers}
\alias{get.TopMarkers}
\title{Select the top marker genes from each cell cluster}
\usage{
get.TopMarkers(
  marker.table,
  gene.col,
  cluster.col,
  fold.change.col,
  p.value.col,
  mito.genes = NULL,
  pval.cutoff = 0.01,
  n.top = 200
)
}
\arguments{
\item{marker.table}{a data.frame or matirx contains gene names, effect size (e.g., fold change), cell cluster and p.value columns.}

\item{gene.col}{the column name of gene names in marker.table.}

\item{cluster.col}{the column name of cell clusters in marker.table.}

\item{fold.change.col}{the column name of fold change in marker.table.}

\item{p.value.col}{the column name of p values marker.table.}

\item{mito.genes}{a vector of mitochondria genes. These genes will be excluded from selected genes. Default is `NULL`.}

\item{pval.cutoff}{a numeric value to specify cutoff of p values.}

\item{n.top}{specify how many genes to be selected from each cell cluster. Default is 200.}
}
\value{
a vector of unique gene names to be used for deconvolution.
}
\description{
This function selects the top marker genes given marker table computated using function `get.Markers.Seurat` or user provided.
The function filters genes given p value cutoff, ranks them according to the absolute log fold change and select top genes.
}
