% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initBetaEst.R
\name{initBetaEst}
\alias{initBetaEst}
\title{Initial estimation of the proportions of cell types}
\usage{
initBetaEst(x, y, w0, w, method = "nnls")
}
\arguments{
\item{x}{the reference data matrix (gene x cell type) of estimated gene expressions (mu)}

\item{y}{a mixture sample to be deconvoluted}

\item{w0}{gene weights related to cell type specificity}

\item{w}{gene weights related to cross sample varibility}

\item{method}{method to perform the initial estimation of the cell proportions. Available choices: 'nnls'(default), 'mle'}
}
\value{
a vector of unnormalized estimated cell type proportions
}
\description{
Initial estimation of the proportions of cell types
}
