% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixCells.R
\name{MixCells}
\alias{MixCells}
\title{Simualte spatial spots using single cell reference data
This function samples cells from single cell reference data and mix them to simulate spatial spots.}
\usage{
MixCells(
  counts,
  annotation,
  celltypes,
  percent,
  batchsize = 10,
  normalize = FALSE,
  nspots = 100,
  replace = F
)
}
\arguments{
\item{counts}{single cell count matrix (gene by cell) where cells are sampled from.}

\item{annotation}{cell type annotation for the cells in the count matrix.}

\item{celltypes}{a vector of cell type names to be included for sampling. The names must exsit in the annotation.}

\item{percent}{a vector of the same length as `celltypes`, specifying the percentages of corresponding cell types to be simulated.}

\item{batchsize}{specify how many cells per spot. Default is 10.}

\item{normalize}{whether to normalize the count matrix before mixing. Default is `FALSE`. If `TRUE`, the cells will be normalized to have total counts of 1e+5.}

\item{nspots}{specify how many spots to be simulated. Default is 100.}

\item{replace}{whether to sample cell with replacement. Defualt is `FALSE`.}
}
\value{
a matrix that simulated spatial spots.
}
\description{
Simualte spatial spots using single cell reference data
This function samples cells from single cell reference data and mix them to simulate spatial spots.
}
