% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin.est.R
\name{negbin.est}
\alias{negbin.est}
\title{Estimate gene mean and dispersion by fitting negative binomial distribution}
\usage{
negbin.est(X)
}
\arguments{
\item{X}{a vector of gene count to fit negative binomial distribution.}
}
\value{
estimated size and mean.
}
\description{
This function takes a vector of gene counts, fits the negative binomial distribution, and estimate the size and mean parameters.
}
