% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.cell.R
\name{normalize.cell}
\alias{normalize.cell}
\title{Normalize single cell count matrix}
\usage{
normalize.cell(counts, annotations, scale.factor = 1e+05, normalize = "Total")
}
\arguments{
\item{counts}{the raw single cell count matrix, sparse matrix (dgCMatrix).}

\item{annotations}{a vector of length equal to the number of cells, each element annotates the cell type of a cell in the count matrix.}

\item{scale.factor}{total number of reads per cell (column) to be scaled to. Default is 1e+5.}

\item{normalize}{it has to one of "median", "mean" and "total".}
}
\value{
a normalized count matrix.
}
\description{
The function normalize the total counts of cells to the medians or means of counts in each cell cluster,
or to the same total number of reads given by `scale.factor`.
}
