% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform.stats.R
\name{perform.stats}
\alias{perform.stats}
\title{Computate the statistics to evaluate deconvolution performance}
\usage{
perform.stats(prop.ests, prop.real, false.cutoff = 1e-04)
}
\arguments{
\item{prop.ests}{a matrix or data frame of estimated proporitons with row being cell types and columns being samples (or spatial spots).}

\item{prop.real}{a matrix or data frame of the ground truth proportions.}

\item{false.cutoff}{a numeric value (between 0 and 1) used to determine whether a cell type is detected. Default value is 1e-4.}
}
\value{
a list of elements including mAD, RMSD, Pearson and Spearman correlations, number of false positives and false negatives.
The function summarizes the overall performance in two ways:
1) the four statisitcs are computated per sample then the means are summarized (stored in `Stat_means`),
and 2) the samples were pooled first then the four statistics are computated over all values (stored in `Stat_all`).
}
\description{
This function computes four evaluation statistics of the estimated proportions given the the ground truth.
The 4 statistics include mean absolute difference (mAD), rooted mean squared deviation (RMSD), Pearson and Spearman correlations.
It also summarizes the number of false postives and false negatives based on a given cutoff of detection.
}
