% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proportion.Barplot.R
\name{Proportion.Barplot}
\alias{Proportion.Barplot}
\title{Make bar plot to display percentages of cell types per sample}
\usage{
Proportion.Barplot(prop.ests)
}
\arguments{
\item{prop.ests}{a matrix or data frame of estimated proporitons with row being cell types and columns being samples (or spatial spots).}
}
\value{
a ggplot object. Use `print` to render the plot.
}
\description{
This function makes bar plot to visualize the percentages of cell types per sample or spatial spots.
}
