% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref.build.R
\name{ref.build}
\alias{ref.build}
\title{Build the single cell reference}
\usage{
ref.build(
  counts,
  annotations,
  genes,
  samples = NULL,
  normalize = "None",
  scale.factor = 1e+05,
  multi.sample.bulk = TRUE,
  multi.sample.single = TRUE,
  silent = FALSE
)
}
\arguments{
\item{counts}{a matrix/data.frame/sparse matrix (dgCMatrix) of single cell counts with the rows being genes and columns being cells.}

\item{annotations}{a vector annotates the cell types of cells.}

\item{genes}{the set of genes selected and to be used for deconvolution.}

\item{samples}{a vector annotates which sample ID the cells come from. Default is `NULL`.}

\item{normalize}{specify how to normalize the single cell raw counts. It should be chosen from "None", "Total" and "Median". Default is "None".}

\item{scale.factor}{the total number of reads to be normalized to if `normalize = "Total"`. Default is 1e+5.}

\item{multi.sample.bulk}{specify whether the bulk to be deconvoluted has multiple samples. Default is TRUE.}

\item{multi.sample.single}{specify whether the single cell reference has multiple samples. Default is TRUE.}

\item{silent}{whether to print out messages. Default is FALSE.}
}
\value{
a list of elements including `mus`, `sizes`, `cell.specificity.w` and `cross.sample.w`.
\itemize{
  \item{mus    }{estimated means for selected genes.}
   \item{sizes    }{estimated size parameters for selected genes.}
   \item{cell.specificity.w    }{cell-type specificity weights.}
   \item{cross.sample.w    }{cross-sample varaibility weights.}
   }
}
\description{
This function takes a vector of gene counts, fits the negative binomial distribution, and estimate the size and mean parameters.
}
