% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC_plot.R
\name{ROC_Plot}
\alias{ROC_Plot}
\title{plot ROC curves to compare two methods}
\usage{
ROC_Plot(
  method1.ests,
  method2.ests,
  prop.real,
  labels = c("Method1", "Method2")
)
}
\arguments{
\item{method1.ests}{a matrix or data frame of estimated proporitons with row being cell types and columns being samples (or spatial spots).}

\item{method2.ests}{a matrix or data frame of estimated proporitons with row being cell types and columns being samples (or spatial spots).}

\item{prop.real}{a matrix or data frame of the ground truth proportions.}

\item{labels}{a vector of two elements labeling the two methods. Default is `c("Method1", "Metthod2")`.}
}
\value{
a ggplot object. Use `print` to render the plot.
}
\description{
This function plots ROC curves to compare two methods given the estimated proportions from each one, and the true proportions.
}
