% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowVars.R
\name{rowVars}
\alias{rowVars}
\title{Compute row-wise variance of a matrix or data.frame}
\usage{
rowVars(X)
}
\arguments{
\item{X}{a matrix or data.frame.}
}
\value{
a vector of row variances
}
\description{
The function computes variance for each row of a matrix or data.frame.
}
