% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize.celltype.R
\name{synthesize.celltype}
\alias{synthesize.celltype}
\title{Synthesize celltype by pooling single cells per cell type per sample}
\usage{
synthesize.celltype(counts, annotations, SampleID)
}
\arguments{
\item{counts}{single cell UMI count matrix or sparse matrix (dgCMatrix), with rownames being genes and colnames being cell names.}

\item{annotations}{cell type annotation, a vector of length equal to the number of columns in counts.}

\item{SampleID}{sample identifications, a vector of length equal to the number of columns in counts.}
}
\value{
simulated cell type specific profile per sample
}
\description{
Synthesize celltype by pooling single cells per cell type per sample
}
